refactor(ui): widgets/screenui を menuframe に改名する
screenui は ui が汎用すぎた。タブ帯・パネル・モーダルの枠を組む責務を名前で明示する。

- internal/widgets/screenui を internal/widgets/menuframe へ。package・import・qualifier・doc を置換
